+++ allegro-client/src/main/java/io/github/mgrtomaszzurawski/allegro/sdk/domain/disputes/model/ChatAuthorRole.java
@@ -38,6 +38,7 @@ public enum ChatAuthorRole {
             case ADMIN -> ADMIN;
             case SYSTEM -> SYSTEM;
             case FULFILLMENT -> FULFILLMENT;
+            default -> UNKNOWN;
         };
     }
 }